Updated Setup button
Removed the Setup Board Button, and moved all of it's creation and functionality to the Custom Board area (instead of a black checker). Setup Board Button now checks to see if any Mech Mats are still in the play area, and moves them off the board before setting up locations and cards.